auth 缺陷修改

+++ guiyu-parent/guiyu-auth-parent/guiyu-auth-dal/src/main/java/com/guiji/user/dao/SysPrivilegeMapper.java
@@ -1,14 +1,12 @@
 package com.guiji.user.dao;
 
-import java.util.List;
-
-import org.apache.ibatis.annotations.Param;
-
 import com.guiji.user.dao.entity.SysPrivilege;
 import com.guiji.user.dao.entity.SysPrivilegeExample;
+import java.util.List;
+import org.apache.ibatis.annotations.Param;
 
 public interface SysPrivilegeMapper {
-    int countByExample(SysPrivilegeExample example);
+    long countByExample(SysPrivilegeExample example);
 
     int deleteByExample(SysPrivilegeExample example);
 
@@ -29,7 +27,6 @@ public interface SysPrivilegeMapper {
     int updateByPrimaryKeySelective(SysPrivilege record);
 
     int updateByPrimaryKey(SysPrivilege record);
-
-	List<Integer> queryOrgIds(@Param("authId") String authId);
-
+    
+    List<Integer> queryOrgIds(@Param("authId") String authId);
 }
